+++ src/testing/models.ts
@@ -1,5 +1,3 @@
-import { TracerEvent } from '../types';
-
 export interface Threshold {
   lt?: number;
   lte?: number;
@@ -21,15 +19,5 @@ export abstract class BaseTestEvaluator<TestCaseType, OutputType> {
   abstract evaluateTestCase(args: {
     testCase: TestCaseType;
     output: OutputType;
-  }): Evaluation | Promise<Evaluation>;
-}
-
-export abstract class BaseEventEvaluator {
-  abstract get id(): string;
-
-  maxConcurrency: number = 10;
-
-  abstract evaluateEvent(args: {
-    event: TracerEvent;
   }): Evaluation | Promise<Evaluation>;
 }
