+++ src/tracer/index.ts
@@ -1,0 +1,7 @@
+export { AutoblocksTracer } from './tracer';
+export {
+  BaseEventEvaluator,
+  Evaluation,
+  type TracerEvent,
+  type SendEventArgs,
+} from './models';
