+++ src/types.ts
@@ -1,13 +1,9 @@
 import { z } from 'zod';
-import { Event } from './client';
-import { BaseEventEvaluator } from './testing/models';
 
 // Use any here so that interfaces can be assigned to this type.
 // https://stackoverflow.com/q/65799316
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type ArbitraryProperties = Record<string | number, any>;
-
-export type TracerEvent = Omit<Event, 'id'>;
 
 export interface PromptTracking {
   id: string;
@@ -18,16 +14,6 @@ export interface PromptTracking {
     template: string;
     properties?: ArbitraryProperties;
   }[];
-}
-
-export interface SendEventArgs {
-  traceId?: string;
-  spanId?: string;
-  parentSpanId?: string;
-  timestamp?: string;
-  properties?: ArbitraryProperties;
-  promptTracking?: PromptTracking;
-  evaluators?: BaseEventEvaluator[];
 }
 
 export interface TimeDelta {
